+++ tests/test-konieczny-bmat8-1.cpp
@@ -167,7 +167,7 @@ namespace libsemigroups {
     Konieczny<BMat8> KS(gens);
     KS.run();
 
-    REQUIRE(KS.cend_regular_D_classes() - KS.cbegin_regular_D_classes()
+    REQUIRE(size_t(KS.cend_regular_D_classes() - KS.cbegin_regular_D_classes())
             == idems.size());
 
     size_t count = 0;
