+++ tests/test-race.cpp
@@ -108,8 +108,8 @@ namespace libsemigroups {
       rc.max_threads(2);
       rc.add_runner(std::make_shared<TestRunner1>());
       rc.add_runner(std::make_shared<TestRunner1>());
-      REQUIRE(rc.end() - rc.begin() == rc.number_runners());
-      REQUIRE(rc.cend() - rc.cbegin() == rc.number_runners());
+      REQUIRE(size_t(rc.end() - rc.begin()) == rc.number_runners());
+      REQUIRE(size_t(rc.cend() - rc.cbegin()) == rc.number_runners());
       REQUIRE(2 == rc.number_runners());
     }
